:new: (IF-18) fixed method name
